+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: `imported` flag from raw private keys into `EdgeWalletInfoFull`
 - added: Integration with change-server.
 
 ## 2.26.2 (2025-03-24)