+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.27.0 (2025-04-28)
+
 - added: Added new Duress Mode feature.
 - added: `imported` flag from raw private keys into `EdgeWalletInfoFull`
 - added: Integration with change-server.